requirements.txt
pypdf>=5.0.0
Pillow>=10.0.0
PyYAML>=6.0.0
pymupdf>=1.24.0
